+++ lib/rley/formatter/json.rb
@@ -5,11 +5,14 @@ module Rley # This module is used as a namespace
   # Namespace dedicated to parse tree formatters.
   module Formatter
 
-    # A formatter class that renders the visit notification events
-    # from a parse tree visitor
-    class Debug < BaseFormatter
+    # A formatter class that renders a parse tree in JSON format
+    class Json < BaseFormatter
       # Current indentation level
       attr(:indentation)
+
+      # Array of booleans (one per indentation level).
+      # Set to true after first child was visited.
+      attr(:sibling_flags)
 
       # Constructor.
       # @param anIO [IO] The output stream to which the rendered grammar
@@ -17,6 +20,7 @@ module Rley # This module is used as a namespace
       def initialize(anIO)
         super(anIO)
         @indentation = 0
+        @sibling_flags = [ false ]
       end
 
       public
@@ -26,7 +30,9 @@ module Rley # This module is used as a namespace
       # parse tree
       # @param _ptree [ParseTree]
       def before_ptree(_ptree)
-        output_event(__method__, indentation)
+        print_text('', "{\n")
+        indent
+        print_text('', "\"root\":")
         indent
       end
 
@@ -34,9 +40,11 @@ module Rley # This module is used as a namespace
       # Notification of a visit event: the visitor is about to visit
       # a non-terminal node
       # @param _nonterm [NonTerminalNode]
-      def before_non_terminal(_nonterm)
-        output_event(__method__, indentation)
-        indent
+      def before_non_terminal(nonterm_node)
+        separator = sibling_flags[-1] ? ",\n" : "\n"
+        name = nonterm_node.symbol.name
+        print_text(separator, "{ \"#{name}\": ")
+        sibling_flags[-1] = true
       end
 
       # Method called by a ParseTreeVisitor to which the formatter subscribed.
@@ -45,34 +53,23 @@ module Rley # This module is used as a namespace
       # @param _parent [NonTerminalNode]
       # @param _children [Array] array of children nodes
       def before_children(_parent, _children)
-        output_event(__method__, indentation)
+        print_text('[', nil)
         indent
+        sibling_flags.push(false)
       end
 
       # Method called by a ParseTreeVisitor to which the formatter subscribed.
       # Notification of a visit event: the visitor is about to visit
       # a terminal node
       # @param _term [TerminalNode]
-      def before_terminal(_term)
-        output_event(__method__, indentation)
+      def before_terminal(term_node)
+        separator = sibling_flags[-1] ? ",\n" : "\n"
+        name = term_node.symbol.name
+        lexeme = term_node.token.lexeme
+        print_text(separator, "{\"#{name}\": \"#{lexeme}\"}")
+        sibling_flags[-1] = true
       end
 
-      # Method called by a ParseTreeVisitor to which the formatter subscribed.
-      # Notification of a visit event: the visitor completed the visit of
-      # a terminal node.
-      # @param _term [TerminalNode]
-      def after_terminal(_term)
-        output_event(__method__, indentation)
-      end
-
-      # Method called by a ParseTreeVisitor to which the formatter subscribed.
-      # Notification of a visit event: the visitor completed the visit of
-      # a non-terminal node
-      # @param _nonterm [NonTerminalNode]
-      def after_non_terminal(_)
-        dedent
-        output_event(__method__, indentation)
-      end
 
       # Method called by a ParseTreeVisitor to which the formatter subscribed.
       # Notification of a visit event: the visitor completed the visit of
@@ -80,8 +77,18 @@ module Rley # This module is used as a namespace
       # @param _parent [NonTerminalNode]
       # @param _children [Array] array of children nodes
       def after_children(_parent, _children)
+        sibling_flags.pop
+        print_text("\n", ']')
         dedent
-        output_event(__method__, indentation)
+        print_text("\n", '}')
+      end
+
+
+      # Method called by a ParseTreeVisitor to which the formatter subscribed.
+      # Notification of a visit event: the visitor completed the visit of
+      # a non-terminal node
+      # @param _nonterm [NonTerminalNode]
+      def after_non_terminal(_)
       end
 
 
@@ -91,7 +98,9 @@ module Rley # This module is used as a namespace
       # @param _ptree [ParseTree]
       def after_ptree(_ptree)
         dedent
-        output_event(__method__, indentation)
+        #print_text("\n", ']')
+        dedent
+        print_text("\n", '}')
       end
 
       private
@@ -104,8 +113,9 @@ module Rley # This module is used as a namespace
         @indentation -= 1
       end
 
-      def output_event(anEvent, indentationLevel)
-        output.puts "#{' ' * 2 * indentationLevel}#{anEvent}"
+      def print_text(aSeparator, aText)
+        output.print aSeparator
+        output.print "#{' ' * 2 * @indentation}#{aText}" unless aText.nil?
       end
 
     end # class